requirements.txt
asgiref==3.3.4
astroid==2.5.6
click==7.1.2
colorama==0.4.4
dig==0.1.10
Django==3.2.3
docopt==0.6.2
futures==3.1.1
gunicorn==20.1.0
h11==0.11.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
psycopg2-binary==2.8.6
pylint==2.8.2
python-decouple==3.4
pytz==2021.1
requests==2.7.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
uvicorn==0.13.3
waitress==1.4.4
watch==0.2.7
wrapt==1.12.1
asgiref==3.3.4
astroid==2.5.6
click==7.1.2
colorama==0.4.4
dig==0.1.10
Django==3.2.3
docopt==0.6.2
futures==3.1.1
gunicorn==20.1.0
h11==0.11.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mysqlclient==2.0.3
psycopg2-binary==2.8.6
pylint==2.8.2
python-decouple==3.4
python-dotenv==0.18.0
pytz==2021.1
requests==2.7.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
uvicorn==0.13.3
waitress==1.4.4
watch==0.2.7
wrapt==1.12.1
